Added assignment for fields.
Note that we only need to support Identifier.Identifier and not an arbitrary number of .field, like a.b.c = 2 or a.b.c.d = 2;
That is because structs and arrays are always references, i.e. pointers. So, the above can be achieved as:
B b = a.b;  // This is a pointer - any change to this will be reflected to a.b
b.c = 2;
For the same reason, we don't need to support something like:
a.b[2] = 3;
because we can achieve that as:
Arr arr = a.b;  // arr is a pointer under the hood - any change to its elements will be reflected to a.b
arr = a.b;