[Lexer] Use int_scan.cpp in lex.cpp
This includes some changes to the previous int scan utils and the lexer.
The major change is that now a lexer is initialized with the input
string which is kept as a global variable inside the lexer.
Also, the handling of the numerical base is different and
some error messages were added.